Move DefaultTripPattern to api package
